# api_gateway/src/main/resources/static/baby-billing-swagger.yaml
openapi: "3.0.0"
info:
  version: 1.0.0
  title: "Baby Billing API"
  license:
    name: MIT
servers:
  - url: http://localhost:8765
...
paths:
...
  /brt/api/subscriber/{msisdn}/changeTariff:
    patch:
      tags:
        - manager
      summary: Смена тарифа абонента
      security:
        - manager_auth: []
      operationId: changeTariff
      description: Позволяет менеджеру сменить тариф абоненту
      parameters:
        - name: msisdn
          in: path
          description: Номер телефона абонента
          required: true
          schema:
            type: integer
            example: 79007770303
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewTariff'
      responses:
        '201':
          description: Тариф успешно изменен
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Subscriber'
        '400':
          $ref: '#/components/responses/InvalidValues'
        '401':
          $ref: '#/components/responses/InvalidAuth'
        '404':
          description: Абонент не найден
        '5XX':
          $ref: '#/components/responses/ServerError'
...
components:
  schemas:
    Subscriber:
      type: object
      properties:
        msisdn:
          type: integer
          description: Номер абонента
          format: int64
          example: 79007770303
        tariffId:
          type: integer
          description: ID тарифа абонента
          format: int32
          example: 01
...
    NewTariff:
      type: object
      properties:
        tariffId:
          type: integer
          description: Новый ID тарифа абонента
          format: int32
          example: 01
  responses:
    ServerError:
      description: Ошибка сервера
    InvalidValues:
      description: Переданы неверные параметры
    InvalidAuth:
      description: Ошибка аутентификации, неверные учетные данные
  securitySchemes:
    manager_auth:
      type: http
      scheme: basic
      description: Авторизация менеджера с указанием логина и пароля
    subscriber_auth:
      type: http
      scheme: basic
      description: Авторизация абонента с указанием логина